+++ resources/migrations/20220823032338-users.up.sql
@@ -1,0 +1,16 @@
+CREATE TABLE users (
+  id INTEGER PRIMARY KEY AUTO_INCREMENT,
+
+  valid BOOLEAN DEFAULT true,
+
+  name    VARCHAR(255) NOT NULL UNIQUE,
+  belong  VARCHAR(255),
+  email   VARCHAR(255),
+
+  cid     VARCHAR(255),
+  secret  VARCHAR(255),
+  access  VARCHAR(255),
+  refresh VARCHAR(255),
+
+  updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+  created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP);